adds copyrights, adds some renaming as well as changed to es6 file exporting syntax
