+++ js/prestashop-ui-kit.js
@@ -3,7 +3,7 @@ import 'bootstrap';
 import 'select2';
 import 'jquery.growl';
 import 'pstagger';
-var initInputFileChangeEvent = require('./events/file-input.js');
+import initInputFile from './components/file-input.js';
 
 $.fn.pstooltip = $.fn.tooltip;
 
@@ -166,7 +166,7 @@ $.fn.pstooltip = $.fn.tooltip;
         initAlerts();
         initToolTips();
         initSearchBar();
-        initInputFileChangeEvent();
+        initInputFile();
       });
     },
